Add anyOf rule
